ml-service/requirements.txt
fastapi
uvicorn
libsql-client
pandas
scikit-learn
matplotlib
seaborn
fpdf2
requests
python-dotenv